% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_country_currency.R
\docType{data}
\name{code_country}
\alias{code_country}
\title{country code list}
\format{A data frame with 249 rows and 4 columns.}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-1}
}
\usage{
data(code_country)
}
\description{
Code list for the representation of names 
of countries and their subdivisions.
}
\examples{
# load code of country
data(code_country)

}
\keyword{data}
