% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_country_currency.R
\docType{data}
\name{code_currency}
\alias{code_currency}
\title{currency code list}
\format{A data frame with 278 rows and 3 columns.}
\source{
\url{https://en.wikipedia.org/wiki/ISO_4217}, \url{https://www.currency-iso.org/en/home/tables/table-a1.html}
}
\usage{
data(code_currency)
}
\description{
Current currency & funds code list.
}
\examples{
# load code of currency
data(code_currency)

}
\keyword{data}
