% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dt_banks}
\alias{dt_banks}
\title{dataset of bank stocks in sse}
\format{
A data frame with 7506 rows and 15 variables:
\describe{
\item{symbol}{stock ticker symbol}
\item{name}{stock ticker name}
\item{date}{trade date}
\item{open}{stock price at the open of trading}
\item{high}{stock price at the highest point during trading}
\item{low}{stock price at the lowest point during trading}
\item{close}{stock price at the close of trading}
\item{volume}{number of shares traded}
\item{amount}{monetary value of shares traded}
\item{turnover}{rate of shares traded over total}
\item{close_adj}{adjusted stock price at the close of trading}
}
}
\usage{
dt_banks
}
\description{
The daily historical data of bank stocks
}
\keyword{datasets}
