% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dt_ssec}
\alias{dt_ssec}
\title{dataset of shanghai composite index}
\format{
A data frame with 7506 rows and 15 variables:
\describe{
\item{symbol}{stock ticker symbol}
\item{name}{stock ticker name}
\item{date}{trade date}
\item{open}{stock price at the open of trading}
\item{high}{stock price at the highest point during trading}
\item{low}{stock price at the lowest point during trading}
\item{close}{stock price at the close of trading}
\item{close_prev}{stock price at the close of previous trading day}
\item{change_pct}{change percentage of stock close price}
\item{volume}{number of shares traded}
\item{amount}{monetary value of shares traded}
\item{turnover}{rate of shares traded over total}
\item{cap_market}{tradable market capitalisation}
\item{cap_total}{total market capitalisation}
\item{close_adj}{adjusted stock price at the close of trading}
\item{unit}{price unit, such as in CNY/USD}
}
}
\usage{
dt_ssec
}
\description{
The daily historical Shanghai Composite Index
}
\keyword{datasets}
