% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_code.R
\name{ed_code}
\alias{ed_code}
\title{code list by category}
\usage{
ed_code(cate = NULL)
}
\arguments{
\item{cate}{The available category values including 'country', 'currency', 'stock_exchange', 'commodity_exchange', 'china_district'.}
}
\description{
\code{ed_code} get the code list of country, currency, stock exchange, commodity exchange and administrative district of mainland of China.
}
\examples{
\dontrun{
# specify the categories
code_list1 = ed_code(cate = c('country', 'currency'))

# interactivly return code list
code_list2 = ed_code()

}

}
