% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_fred.R
\name{ed_fred}
\alias{ed_fred}
\title{get economic data from FRED}
\usage{
ed_fred(symbol = NULL, date_range = "10y", from = NULL,
  to = Sys.Date(), na_rm = TRUE, print_step = 1L)
}
\arguments{
\item{symbol}{symbol of indicators in FRED, which is available via function ed_FRED_symbol or its website.}

\item{from}{the start date. Default is '1776-07-04'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
ed_fred provides an interface to access the economic data provided by FRED (\url{https://fred.stlouisfed.org})
}
\examples{
\dontrun{
dat = ed_fred(c("A191RL1A225NBEA", "GDPCA"))

}

}
