% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_fred.R
\name{ed_fred_symbol}
\alias{ed_fred_symbol}
\title{query symbols of economic data in FRED}
\usage{
ed_fred_symbol(keywords = NULL)
}
\arguments{
\item{keywords}{the query text. If it is NULL, the function will search symbols by category.}
}
\description{
ed_fred_symbol provides an interface to search symbols of economic data in FRED by keywords or category.
}
\examples{
\dontrun{
# search data by keywords
sybs1 = ed_fred_symbol("gdp china")

# search data by category
sybs1 = ed_fred_symbol()
}

}
