% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_nbs.R
\name{ed_nbs_region}
\alias{ed_nbs_region}
\title{query code of subregion in China from NBS}
\usage{
ed_nbs_region(geo_type = "", eng = FALSE)
}
\arguments{
\item{geo_type}{geography type in NBS, including 'province', 'city'. Default is "".}

\item{eng}{logical. Default is FALSE. If it is FALSE, the result is in Chinese, otherwise in English.}
}
\description{
\code{ed_nbs_region} query province or city code from NBS
}
\examples{
#  province code 
prov1 = ed_nbs_region(geo_type = 'province') 
# or using 'p' represents 'province'
prov2 = ed_nbs_region(geo_type = 'p') 

\dontrun{
#  city code in Chinese
city = ed_nbs_region(geo_type = 'c', eng = FALSE) 
#  city code in English
city = ed_nbs_region(geo_type = 'c', eng = TRUE) 
}

}
