% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_exchange_commodity.R
\docType{data}
\name{exchange_commodity}
\alias{exchange_commodity}
\title{commodity exchanges list}
\format{A data frame with 97 rows and 7 columns.}
\source{
\url{https://en.wikipedia.org/wiki/List_of_commodities_exchanges}
}
\usage{
data(exchange_commodity)
}
\description{
List of commodities exchanges
}
\examples{
# load commodity exchange
data(exchange_commodity)

}
\keyword{data}
