% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_exchange_stock.R
\docType{data}
\name{exchange_stock}
\alias{exchange_stock}
\title{Stock Exchanges}
\format{A data frame with 65 rows and 13 columns.}
\source{
\url{https://en.wikipedia.org/wiki/List_of_stock_exchange_trading_hours}
}
\usage{
data(exchange_stock)
}
\description{
List of major stock exchanges by January 2015 (top 20 by market capitalization)
}
\examples{
# load stock exchange
data(exchange_stock)

}
\keyword{data}
