% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geted_fred.R
\name{geted_fred}
\alias{geted_fred}
\title{get economic data from FRED}
\usage{
geted_fred(symbol, from = "1776-07-04", to = "9999-12-31",
  print_step = 1L)
}
\arguments{
\item{symbol}{symbol of indicators in FRED, which is available via function geted_FRED_symbol or its website.}

\item{from}{the start date. Default is '1776-07-04'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
geted_fred provides an interface to access the economic data provided by FRED (\url{https://fred.stlouisfed.org})
}
\examples{
\dontrun{
dat = geted_fred(c("A191RL1A225NBEA", "GDPCA"))

}

}
