% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geted_nbs.R
\name{geted_nbs}
\alias{geted_nbs}
\title{get economic data from National Bureau of Statistics of China (NBS)}
\source{
\url{http://data.stats.gov.cn/index.htm}
}
\usage{
geted_nbs(geo_type = NULL, frequency = NULL, symbol = NULL,
  region = NULL, from = "2010-01-01", to = Sys.Date(), na_rm = FALSE,
  eng = FALSE)
}
\arguments{
\item{geo_type}{geography type in NBS, including 'national', 'province', 'city'. Default is NULL.}

\item{frequency}{the frequency of indicators in NBS, including 'monthly', 'quarterly', 'yearly'. Default is NULL.}

\item{symbol}{symbol of indicators in NBS, which is available via geted_nbs_symbol. Default is NULL.}

\item{region}{region codes of province or city, which is available via geted_nbs_region. Default is NULL.}

\item{from}{the start date. Default is '2010-01-01'.}

\item{to}{the end date. Default is current system date.}

\item{na_rm}{logical. If it is TRUE, the missing values will be removed. Default is FALSE.}

\item{eng}{logical. If it is FALSE, the query results are in Chinese, otherwise in English. Default is FALSE.}
}
\description{
\code{geted_nbs} provides an interface to get economic data from National Bureau of Statistics of China (NBS).
}
\examples{
# interactively setting paratmeters
\dontrun{
dt = geted_nbs()

# specify paratmeters
dt1 = geted_nbs(geo_type='national', frequency='quarterly', 
  symbol='A010101')

# or using 'n'/'q' represents 'national'/'quarterly'
dt2 = geted_nbs(geo_type='n', frequency='q', symbol='A010101')


# get data of one region
dt3 = geted_nbs(geo_type='province', frequency='quarterly', 
  symbol='A010101', region='110000', from='2010-03-01', to='2010-03-01')
  
# get data of all province
dt4 = geted_nbs(geo_type='province', frequency='quarterly', 
  symbol='A010101', region='all', from='2010-03-01', to='2010-03-01')
}

}
