% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geted_nbs.R
\name{geted_nbs_region}
\alias{geted_nbs_region}
\title{get code of subregion in China from NBS}
\usage{
geted_nbs_region(geo_type = "", eng = FALSE)
}
\arguments{
\item{geo_type}{geography type in NBS, including 'province', 'city'. Default is "".}

\item{eng}{logical. Default is FALSE. If it is FALSE, the result is in Chinese, otherwise in English.}
}
\description{
\code{geted_nbs_region} get province or city code from NBS
}
\examples{
# get province code 
prov1 = geted_nbs_region(geo_type = 'province') 
# or using 'p' represents 'province'
prov2 = geted_nbs_region(geo_type = 'p') 

\dontrun{
# get city code in Chinese
city = geted_nbs_region(geo_type = 'c', eng = FALSE) 
# get city code in English
city = geted_nbs_region(geo_type = 'c', eng = TRUE) 
}

}
