% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfd.R
\name{getfd_163}
\alias{getfd_163}
\title{get financial statement data from 163}
\usage{
getfd_163(symbol, type = NULL)
}
\arguments{
\item{symbol}{symbol of Chinese stock in 163}

\item{type}{the type of financial statement table}
}
\description{
`getfd_163` gets the financial statement tables of Chinese stock from 163.
}
\examples{
\\dotrun{
dat1 = getfd_163("000001")

dat2 = getfd_163("000001", type="fs0")

dat3 = getfd_163("000001", type="fs0_summary")

}

}
