% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfd.R
\name{getfd_cn}
\alias{getfd_cn}
\title{get financial statement data of Chinese stock}
\usage{
getfd_cn(symbol, type = NULL)
}
\arguments{
\item{symbol}{symbol of Chinese stock}

\item{type}{the type of financial statement table}
}
\description{
`getfd_cn` gets the financial statement tables of Chinese stock.
}
\examples{
\\dotrun{
dat1 = getfd_cn("000001")

dat2 = getfd_cn("000001", type="fs0")

dat3 = getfd_cn("000001", type="fs0_summary")

}

}
