% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfr_cn.R
\name{getfr}
\alias{getfr}
\title{get financial reports and indicators}
\usage{
getfr(symbol, type = NULL, print_step = 1L, region = "cn")
}
\arguments{
\item{symbol}{symbol of stock in China.}

\item{type}{the type of financial reports.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}

\item{region}{only cn (China) is available.}
}
\description{
\code{getfr} provides an interface to get financial reports and indicators of listed companies in China.
}
\examples{
\dontrun{
# interactively specify type of financial table 
dat1 = getfr("000001")

# manually specify type of financial table
# type = "fr0"
dat2 = getfr("000001", type="fr0")
# or type = "fr0_summary"
dat3 = getfr("000001", type="fr0_summary")

# multiple symbols and reports
dat4 = getfr(c("000001", "600000"), type = "fi")

}

}
