% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfs_cn.R
\name{getfs}
\alias{getfs}
\title{get financial statements and indicators}
\usage{
getfs(symbol, type = NULL, print_step = 1L, region = "cn")
}
\arguments{
\item{symbol}{symbol of stock in China.}

\item{type}{the type of financial statements.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}

\item{region}{only cn (China) is available.}
}
\description{
\code{getfs} provides an interface to get financial statements and indicators of listed companies in China.
}
\examples{
\dontrun{
# interactively specify type of financial table 
dat1 = getfs("000001")

# manually specify type of financial table
# type = "fs0"
dat2 = getfs("000001", type="fs0")
# or type = "fs0_summary"
dat3 = getfs("000001", type="fs0_summary")

# multiple symbols and reports
dat4 = getfs(c("000001", "600000"), type = "fi")

}

}
