% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd.R
\name{getmd}
\alias{getmd}
\title{get market data}
\usage{
getmd(symbol, market = NULL, exchange = NULL, source = "yahoo",
  frequency = "daily", from = "1900-01-01", to = Sys.time(),
  print_step = 1L)
}
\description{
query market data from public sources.
}
