% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_163.R
\name{getmd_163}
\alias{getmd_163}
\title{Get Stock Market Data from 163}
\source{
\url{http://quotes.money.163.com/stock}
}
\usage{
getmd_163(symbol, from = "1900-01-01", to = Sys.Date(), print_step = 1,
  frequency = "daily", fillzero = FALSE)
}
\arguments{
\item{symbol}{symbol of Chinese stock and index in 163, which should be in end of '.ss' or '.sz', represents sse and szse, such as '600000.ss', '000001.sz'.}

\item{from}{the start date. Default is '1900-01-01'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print variable names by each print_step-th iteration. Default is 1.}

\item{frequency}{currently only daily data is available.}

\item{fillzero}{logical. Defualt is TRUE. If it is TRUE, the zeros in dataset will be filled with last non-zero values.}
}
\description{
\code{getmd_163} gets Chinese stock and index data from 163, and returns a list of dataframes.
}
\examples{
\dontrun{
dat = getmd_163(symbol=c('600000', '000001'))

ssec_szsec = getmd_163(symbol = c('^000001', '^399001'))
}

}
