% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_163.R
\name{getmd_163_symbol}
\alias{getmd_163_symbol}
\title{Get Stock and index Symbols from 163}
\source{
\url{http://quotes.money.163.com/stock}
}
\usage{
getmd_163_symbol()
}
\description{
\code{getmd_163_symbol} gets all Chinese stock and index symbols from 163.
}
\examples{
\dontrun{
dt = getmd_163_symbol()}

}
