% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_commodity_cn.R
\name{getmd_commodity}
\alias{getmd_commodity}
\title{get commodity market data}
\usage{
getmd_commodity(symbol, print_step = 1, region = "cn")
}
\arguments{
\item{symbol}{symbols of Chinese commodity future. For more commodity symbols go to \url{http://vip.stock.finance.sina.com.cn/quotes_service/view/qihuohangqing.html}.}

\item{print_step}{A non-negative integer, which will print variable names by each print_step-th iteration. Default is 1.}

\item{region}{only cn (China) is available.}
}
\description{
\code{getmd_commodity} gets Chinese commodity future data.
}
\examples{
\dontrun{
dat = getmd_commodity(symbol='RB0')

dat = getmd_commodity(symbol=c("AG0","AU0"))
}

}
