% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_cons_csi.R
\name{getmd_cons_csi}
\alias{getmd_cons_csi}
\title{Get constituent of China securities index}
\source{
\url{http://www.csindex.com.cn/zh-CN}
}
\usage{
getmd_cons_csi(symbol)
}
\arguments{
\item{symbol}{the code of China securities index.}
}
\description{
\code{getmd_csi_cons} provides an API to get the current constituent of China securities index.
}
\examples{
\dontrun{
dt50 = getmd_cons_csi("000016")

dt300 = getmd_cons_csi("000300")

dt500 = getmd_cons_csi("000905")
}

}
