% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_csi_cons.R
\name{getmd_index_cons}
\alias{getmd_index_cons}
\title{get constituent of China securities index (CSI).}
\source{
\url{http://www.csindex.com.cn/zh-CN}
}
\usage{
getmd_index_cons(symbol)
}
\arguments{
\item{symbol}{the code of China securities index.}
}
\description{
\code{getmd_index_cons} provides an API to get the current constituent of China securities index.
}
\examples{
\dontrun{
dt50 = getmd_index_cons("000016")

dt300 = getmd_index_cons("000300")

dt500 = getmd_index_cons("000905")
}

}
