% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_stock_cn.R
\name{getmd_stock}
\alias{getmd_stock}
\title{get stock market data}
\usage{
getmd_stock(symbol, from = "1900-01-01", to = Sys.Date(), print_step = 1,
  frequency = "daily", fillzero = FALSE, region = "cn")
}
\arguments{
\item{symbol}{symbols of Chinese stock and index, the index symbol should starts with ^.}

\item{from}{the start date. Default is '1900-01-01'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print variable names by each print_step-th iteration. Default is 1.}

\item{frequency}{supports daily and spot data.}

\item{fillzero}{logical. Defualt is FALSE If it is TRUE, the zeros in dataset will be filled with last non-zero values.}

\item{region}{only cn (China) is available.}
}
\description{
\code{getmd_stock} provides an interface to get end of date and spot price data of Chinese stock and index.
}
\examples{
\dontrun{
dat = getmd_stock(symbol=c('600000', '000001', '^000001', '^399001'))

dat2 = getmd_stock(symbol=c('600000', '000001', '^000001', '^399001'), frequency="spot")

# get spot price of all A shares
dat3 = getmd_stock(symbol='a', frequency="spot")

# get spot price of all index in sse and szse
dat4 = getmd_stock(symbol='index', frequency="spot")
}

}
