% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_stock_cn.R
\name{getmd_stock_cn}
\alias{getmd_stock_cn}
\title{get stock market data in sse and szse}
\usage{
getmd_stock_cn(symbol, from = "1900-01-01", to = Sys.Date(),
  print_step = 1, frequency = "daily", fillzero = FALSE)
}
\arguments{
\item{symbol}{symbols of Chinese stock and index, the index symbol should starts with ^.}

\item{from}{the start date. Default is '1900-01-01'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print variable names by each print_step-th iteration. Default is 1.}

\item{frequency}{supports daily and spot data.}

\item{fillzero}{logical. Defualt is FALSE If it is TRUE, the zeros in dataset will be filled with last non-zero values.}
}
\description{
\code{getmd_stock_cn} gets Chinese stock and index data in sse and szse. The return object is a list of dataframes.
}
\examples{
\dontrun{
dat = getmd_stock_cn(symbol=c('600000', '000001'))

ssec_szsec = getmd_stock_cn(symbol = c('^000001', '^399001'))
}

}
