% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_stock_symbol.R
\name{getmd_stock_symbol}
\alias{getmd_stock_symbol}
\title{get stock and index symbols}
\usage{
getmd_stock_symbol(exchanges = c("sse", "szse"))
}
\arguments{
\item{exchanges}{default is c("sse", "szse"), which means to return symbols in sse and szse sse and szse.}
}
\description{
\code{getmd_163_symbol} gets the share symbols in sse (Shanghai Stock Exchange), szse (Shenzhen Stock Exchange) and hkex (Hong Kong Stock Exchange).
}
\examples{
\dontrun{
dt = getmd_stock_symbol()}

}
