% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd.R
\name{getmd_symbol}
\alias{getmd_symbol}
\title{get symbols of market data}
\usage{
getmd_symbol(market = NULL, source = NULL)
}
\arguments{
\item{market}{there are six markets, 'currency', 'bond', 'money', 'index', 'stock' and 'commodity'.}

\item{source}{data sources. The available data sources including 'yahoo', '163', 'sina' and 'original'.}
}
\description{
\code{getmd_symbol} provides an interface to get symbols of different markets, which can query data via \code{getmd}.
}
\examples{
\dontrun{
# specify source
syb_163 = getmd_symbol(source="163")

# specify market
syb_stock = getmd_symbol(market="stock")

# specify both source and market
syb_commodity_sina = getmd_symbol(market="commodity", source="sina")

# without specify both source and market
sybs = getmd_symbol()
}

}
