% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_stock_symbol.R
\name{getmd_symbol_163}
\alias{getmd_symbol_163}
\title{get stock and index symbols}
\usage{
getmd_symbol_163(exchanges = c("sse", "szse"))
}
\arguments{
\item{exchanges}{name of stock exchange. Accepted values including sse, szse and hkex. Default is c("sse", "szse").}
}
\description{
\code{getmd_163_symbol} gets the stock and index symbols in sse (Shanghai Stock Exchange), szse (Shenzhen Stock Exchange) and hkex (Hong Kong Stock Exchange).
}
\examples{
\dontrun{
dt = getmd_stock_symbol()}

}
