% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_commodity_cn.R
\name{getmd_symbol_sina}
\alias{getmd_symbol_sina}
\title{get commodity symbols}
\usage{
getmd_symbol_sina()
}
\description{
\code{getmd_commodity_symbol} gets the symbols commodity future in Dalian Commodity Exchange (dce), Shanghai Futures Exchange (shfe), Shanghai Gold Exchange (sge), Zhengzhou Commodity Exchange (zce). For more commodity symbols go to \url{http://vip.stock.finance.sina.com.cn/quotes_service/view/qihuohangqing.html}.
}
\examples{
\dontrun{
dt = getmd_commodity_symbol()}

}
