% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_opr.R
\name{\%()\%}
\alias{\%()\%}
\alias{\%[)\%}
\alias{\%(]\%}
\alias{\%[]\%}
\title{interval operators}
\usage{
x \%()\% rng

x \%[)\% rng

x \%(]\% rng

x \%[]\% rng
}
\arguments{
\item{x}{numeric vectors}

\item{rng}{numeric vectors, top and bottom limitation}
}
\description{
Binary operators which create the interval signals.
}
\examples{
# 0:4 \%()\% c(1,3)

# 0:4 \%[)\% c(1,3)

# 0:4 \%(]\% c(1,3)

# 0:4 \%[]\% c(1,3)

}
