% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_0cate.R
\name{md_cate}
\alias{md_cate}
\title{main market data by category}
\usage{
md_cate(cate = NULL, symbol = NULL, date_range = "3y", from = NULL,
  to = Sys.Date(), print_step = 1L, ...)
}
\description{
main market data by category
}
