% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_future.R
\name{md_future}
\alias{md_future}
\title{get future data}
\usage{
md_future(symbol = NULL, freq = "daily", date_range = "3y",
  from = NULL, to = Sys.Date(), print_step = 1L)
}
\description{
get future data
}
