% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_future.R
\name{md_future_symbol}
\alias{md_future_symbol}
\title{symbol of future market data}
\usage{
md_future_symbol()
}
\description{
\code{md_future_symbol} search the symbols in future market indicators that provided by sina finance only currently.
}
\examples{
\dontrun{
# interactivly search future market symbols
sybs = md_future_symbol()

}

}
