% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_2money.R
\name{md_money}
\alias{md_money}
\title{query interbank offerd rate}
\usage{
md_money(symbol = NULL, type = "history", date_range = "3y",
  from = NULL, to = Sys.Date(), print_step = 1L)
}
\arguments{
\item{symbol}{ibor symbols. Default is NULL.}

\item{type}{the data type. Default is history.}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is 3y.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{print_step}{a non-negative integer, which will print symbol name by each print_step iteration. Default is 1L.}
}
\description{
\code{md_money} query libor from FRED or shibor from chinamoney.
}
