% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_2chinamoney.R
\name{md_moneycn}
\alias{md_moneycn}
\title{query chinese benchmark rates}
\usage{
md_moneycn(symbol = NULL, date_range = "3y", from = NULL,
  to = Sys.Date(), print_step = 1L)
}
\arguments{
\item{symbol}{benchmarks, available values including 'rmbx', 'shibor', 'lpr', 'pr', 'yb'. Default is NULL,}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is 3y.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{print_step}{a non-negative integer, which will print symbol name by each print_step iteration. Default is 1L.}
}
\description{
\code{md_moneycn} query benchmark rates from chinamoney.com.cn.
}
