% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_stock1_.R
\name{md_stock}
\alias{md_stock}
\title{get stock market data}
\usage{
md_stock(symbol, source = "yahoo", freq = "daily", date_range = "3y",
  from = NULL, to = Sys.Date(), type = "history", adjust = FALSE,
  print_step = 1L, ...)
}
\arguments{
\item{symbol}{symbols stock share.}

\item{source}{data source. The available sources including 'yahoo' and '163'.}

\item{freq}{default is daily. It supports daily, weekly and monthly for data from yahoo; daily for data from 163.}

\item{date_range}{date range. Available value including '1m'-'11m', 'ytd', 'max' and '1y'-. Default is '3y'.}

\item{from}{the start date. Default is NULL.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer. Print symbol name by each print_step iteration. Default is 1L.}

\item{...}{Additional parameters.}
}
\description{
\code{md_stock} provides an interface to query historical stock data.
}
\examples{
\dontrun{
# Example I
# query Chinese stock data from 163
dat1 = md(symbol=c('600000', '000001', '^000001', '^399001'), 
             source="163")

dat2 = md(symbol=c('600000', '000001', '^000001', '^399001'), 
             source="163", freq="spot")

# query spot price of all A shares in sse and szse
dat3 = md(symbol='a', source="163", freq="spot")

# query spot price of all index in sse and szse
dat4 = md(symbol='index', source="163", freq="spot")


# Example II
# query Chinese commodity future price
dat_com1 = md(symbol='RB0', source="sina")

dat_com2 = md(symbol=c("AG0","AU0"), source="sina")


 # Example III
 # query data from yahoo
 dat_yahoo = md(symbol=c("^GSPC", "000001.SS", "EURUSD=X"))
 
 # for Chinese share
 dat_yahoo2 = md(c("000001", "^000001"))
}

}
