% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_0query.R
\name{md_symbol}
\alias{md_symbol}
\title{symbol of market data by category}
\usage{
md_symbol(market = NULL, ...)
}
\arguments{
\item{market}{the market category, including forex, money, bond, stock, future. Default is NULL.}

\item{...}{ignored parameters}
}
\description{
\code{md_stock_symbol} returns all symbols by market category, including forex, money, bond, stock, future.
}
\examples{
\dontrun{
syblst = md_symbol()
}

}
