% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_0pd_code.R
\name{pd_code}
\alias{pd_code}
\title{code list by category}
\usage{
pd_code(cate = NULL)
}
\arguments{
\item{cate}{The available category values including 'country', 'currency', 'stock_exchange', 'commodity_exchange'.}
}
\description{
\code{pd_code} get the code list of country, currency, stock exchange and commodity exchange.
}
\examples{
\donttest{
# specify the categories
code_list1 = pd_code(cate = c('country', 'currency'))

# interactivly return code list
code_list2 = pd_code()

}

}
