% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_index.R
\name{pd_index}
\alias{pd_index}
\title{create an index}
\usage{
pd_index(dt, chain_index, weight, base_index = 1,
  base_date = "2010-01-01", name = NULL)
}
\arguments{
\item{dt}{input timeseries datasets}

\item{chain_index}{the name of chain index}

\item{weight}{the name of weight variable}

\item{base_index}{the base value of index, default is 1}

\item{base_date}{the date base of index, default is "2010-01-01"}

\item{name}{the index name, default is NULL}
}
\description{
pd_index using the method of geometrically weighted averages to create an index based on multiple timeseries datasets.
}
\examples{
\dontrun{
# example I bank share index
banks_symbol = c("601988", "601288", "601398", "601939", "601328")

dat = getmd(banks_symbol, source="163", from="1900-01-01")

dat = lapply(dat, function(x) {
x$change_pct = x$change_pct/100+1
return(x)})

bankindex = pd_index(dat, chain_index="change_pct", weight="cap_total")

# example II golden share index
# sybs = getmd_symbol(market = "stock", source="163")
# sybs_sub = sybs[sector == "有色金属采选" & grepl("黄金|金", name)]

gold_share = c("002155", "002237", "600146", "600311", "600489", "600547", "600687", "600766", "600988", "601069", "601899", "601958")

dat = getmd(gold_share, source="163", from="1900-01-01")

dat = lapply(dat, function(x) {
  x$change_pct = x$change_pct/100+1
  return(x)})

gold_index = pd_index(dat, chain_index="change_pct", weight="cap_total")


# III
securities = sybs[sector == '证券' & !is.na(board), symbol]

dat = getmd(securities, source="163", date_range = 'max')
dat = lapply(dat, function(x) {
  x$change_pct = x$change_pct/100+1
  return(x)})

securities_index = pd_index(dat, chain_index="change_pct", weight="cap_total")

# pd_plot(securities_index)

}

}
