% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_perf.R
\name{pd_perf}
\alias{pd_perf}
\title{create performance of data sets}
\usage{
pd_perf(dt, date_range = "max", from = NULL, to = Sys.Date(),
  x = "close|value")
}
\description{
create performance of data sets
}
