% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_to_freq.R
\name{pd_to_freq}
\alias{pd_to_freq}
\title{convert the frequency of daily data}
\usage{
pd_to_freq(dt, freq, print_step = 0L)
}
\arguments{
\item{dt}{time series datasets}

\item{freq}{the frequency that the input data will converted to. It supports weekly, monthly, quarterly and yearly.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
convert the frequency of daily data
}
\examples{
dts = md_stock(c("^000001", "000001"), date_range = 'max', source = '163')

dts_weekly = pd_to_freq(dts, "weekly")

}
