% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_ti.R
\name{ped_addti}
\alias{ped_addti}
\title{create technical indicators}
\usage{
ped_addti(dt, ti = list(sma = 20, sma = 50), print_step = 1L)
}
\arguments{
\item{dt}{time series datasets}

\item{ti}{list of technical indicators, overlay indicators include mm, sma, ema, smma, bb, sar, and oscillators indicators such as macd, ppo, roc, rsi, cci.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
\code{ped_addti} adds technical indicators to a dataset
}
\examples{
dt = getmd("^000001", source=163)

dt_ti = ped_addti(dt, ti=list(sma=20, sma=50))

}
