% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_dailyto.R
\name{ped_dailyto}
\alias{ped_dailyto}
\title{convert the frequency of daily data}
\usage{
ped_dailyto(dt, freq, print_step = 0L)
}
\arguments{
\item{dt}{time series datasets}

\item{freq}{freq the frequency that the input data will converted to. It supports weekly, monthly, quarterly and yearly.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
convert the frequency of daily data
}
\examples{
dts = getmd(c("^000001", "000001"), from = "1990-01-01", source = "163")

dts_weekly = ped_dailyto(dts, "weekly")

}
