% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_index.R
\name{ped_index}
\alias{ped_index}
\title{create a index}
\usage{
ped_index(dt, chain_index, weight, base_index = 1, base_date = "2010-01-01")
}
\arguments{
\item{dt}{input datasets including many dataframes.}

\item{chain_index}{the name of chain index}

\item{weight}{the name of weight variable}

\item{base_index}{the base value of index, default is 1}

\item{base_date}{the date base of index, deafult is "2010-01-01"}
}
\description{
ped_index creates a index, which is geometrically weighted averages of chain index, based on multiple time series dataframes.
}
\examples{
\dontrun{
banks_symbol = c("601988", "601288", "601398", "601939")

dat = getmd_163(banks_symbol)

library(data.table)
dat = lapply(dat, function(x) x[, pct_change:=pct_change+1])

bankindex = ped_index(dat, "pct_change", "total_cap")
}

}
