% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_plot.R
\name{ped_plot}
\alias{ped_plot}
\title{create a chart for timeseries data}
\usage{
ped_plot(dt, y = "close|value", date_range = "max", from = NULL,
  to = Sys.Date(), title = NULL, type = "line", yaxis_log = FALSE,
  perf = FALSE, multi_series = FALSE, linear_trend = FALSE)
}
\arguments{
\item{dt}{a time series dataset}

\item{y}{the variable display on chart}

\item{date_range}{date range. Available value including YTD, Max and integers from 1 to 1000. Default is Max.}

\item{from}{the start date. Default is max date in input data.}

\item{to}{the end date. Default is min date in data.}

\item{title}{chart title. If it is not specified, the symbol of dataset will be used as chart title.}

\item{type}{chart type, including line, step, candle.}

\item{yaxis_log}{logical, default is FALSE.}

\item{perf}{logical, whether to show the performance of input data. Default is FALSE.}

\item{multi_series}{logical, whether to show multiple series. Default is FALSE.}

\item{linear_trend}{non-negative numeric. Default is 0, which donot show linear trend line. If it larger than 0, show linear_trend*sd from linear trend line.}
}
\description{
`ped_plot` creates a charts for a time series dataset.
}
\examples{
\dontrun{
ssec = getmd("^000001", source="163")

p = ped_plot(ssec, title="SSEC")
print(p)
}

}
