% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_ti.R
\name{ped_ti}
\alias{ped_ti}
\title{create technical indicators}
\usage{
ped_ti(dt, ti = list(sma = list(n = 20), sma = list(n = 50)), ...)
}
\arguments{
\item{dt}{time series datasets}

\item{ti}{list of technical indicators, overlay indicators include mm, sma, ema, smma, bb, sar, and oscillators indicators such as macd, roc, ppo, rsi, cci.}

\item{...}{ignored}
}
\description{
\code{ped_ti} adds technical indicators to a dataset
}
\examples{
dt = getmd("^000001", source=163)

dt_ti = ped_ti(dt, ti=list(sma=20, sma=50))

}
