% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_addti.R
\name{pq_addti}
\alias{pq_addti}
\title{create technical indicators}
\usage{
pq_addti(dt, ...)
}
\arguments{
\item{dt}{time series datasets}

\item{...}{ignored}

\item{ti}{list of technical indicators, overlay indicators include mm, sma, ema, smma, bb, sar, and oscillators indicators such as macd, roc, ppo, rsi, cci.
\enumerate{
\item overlay technical indicators
}

\itemize{
\item sma
\item ema
\item smma
\item bb
\item sar
}
\enumerate{
\item oscillator technical indicators
}

\itemize{
\item macd
\item ppo
\item rsi
\item cci
\item roc
}}
}
\description{
\code{pq_addti} adds technical indicators to a dataset
}
\examples{
dt = md_stock("^000001", source='163')

dt_ti = pq_addti(dt, SMA=list(n=20), SMA=list(n=50))

}
