% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_backtest.R
\name{pq_backtest}
\alias{pq_backtest}
\title{Strategy backtesting}
\usage{
pq_backtest(dt, addti = NULL, init_equity = NULL, date_range = "max",
  from = NULL, to = Sys.Date(), rule = list(long = list(enter = NULL,
  exit = NULL, price = NULL, stop_limit = NULL, position = NULL)),
  stp_lmt_pct = 0.05, show_plot = TRUE, ...)
}
\arguments{
\item{dt}{a list/dataframe of time series dataset}

\item{addti}{list of technical indicators or numerical columes in dt. For technical indicator, it is calculated via \code{pq_addti}, which including overlay and oscillator indicators.}

\item{init_equity}{initial equity}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{rule}{rules of trade strategy.}

\item{stp_lmt_pct}{stop limit percent}

\item{show_plot}{show plot}

\item{...}{Additional parameters}
}
\description{
\code{pq_backtest} provides a simple way to backtest a trade strategy.
}
