% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_perf.R
\name{pq_perf}
\alias{pq_perf}
\title{create performance of data sets}
\usage{
pq_perf(dt, date_range = "max", from = NULL, to = Sys.Date(),
  x = "close|value")
}
\description{
create performance of data sets
}
