% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_performance.R
\name{pq_performance}
\alias{pq_performance}
\title{calculating performance metrics}
\usage{
pq_performance(dt, Ra, Rb = NULL, perf_fun, ...)
}
\arguments{
\item{dt}{a list/dataframe of time series datasets.}

\item{Ra}{the column name of asset returns.}

\item{Rb}{the column name of baseline returns, defaults to NULL.}

\item{perf_fun}{performance function from \code{PerformanceAnalytics} package, see \code{pq_perf_funs}.}

\item{...}{additional parameters, the arguments used in \code{PerformanceAnalytics} functions.}
}
\description{
\code{pq_performance} calculates performance metrics based on returns of market price or portfolio. The performance analysis functions are calling from \code{PerformanceAnalytics} package, which includes many widely used performance metrics.
}
\examples{
 
\donttest{
library(data.table)
library(pedquant) 

# load data
data(ssec)
# transfer prices to returns
datret = pq_return(ssec, 'close', freq = 'monthly', rcol_name = 'Ra')
datret2 = rbindlist(datret)[, date := as.Date(sub('[0-9]{2}$', '01', date))][]

# merge returns of assets and baseline
datRaRb = merge(
    datret2[symbol != '000001.SS', .(date, symbol, Ra)], 
    datret2[symbol == '000001.SS', .(date, Rb = Ra)],
    by = 'date', all.x = TRUE
)

# claculate table.CAPM metrics
perf_capm = pq_performance(datRaRb, Ra = 'Ra', Rb = 'Rb', perf_fun = 'table.CAPM')
rbindlist(perf_capm, idcol = 'symbol')
}
}
