% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_performance.R
\name{pq_performance_funs}
\alias{pq_performance_funs}
\title{performance functions}
\usage{
pq_performance_funs()
}
\description{
A complete list of performance functions from \code{PerformanceAnalytics} package.
}
