% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_perf.R
\name{pq_portfolio}
\alias{pq_portfolio}
\title{creating the equity trends for portfolio}
\usage{
pq_portfolio(dt, w, init_equity = NULL, date_range = "max",
  from = NULL, to = Sys.Date(), x = "close|value",
  summarise = FALSE)
}
\arguments{
\item{dt}{a list/dataframe of time series dataset}

\item{w}{a dataframe of position and price of all transactions.}

\item{init_equity}{initial equity}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{x}{the name of column to calculate. Default is 'close|value'.}

\item{summarise}{logical, whether to display summary of portfolio position}
}
\description{
\code{pq_portfolio} creates the equity trends of a portfolio secturites.
}
\examples{
\donttest{
dat = md_stock(c('FB', 'AMZN', 'AAPL', 'NFLX', 'GOOG'), date_range = 'max')
w = data.frame(
      date = c('2016-01-08'),
      symbol = c('FB', 'AMZN', 'AAPL', 'NFLX', 'GOOG'), 
      price = c(99.88, 618.88, 97.88, 115.88, 716.88),
      position = c(100, 200, 300, 400, 500)
    )
pf = pq_portfolio(dat, w)
}
}
