% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_return.R
\name{pq_return}
\alias{pq_return}
\title{Returns}
\usage{
pq_return(dt, x = "close|value", type = "arithmetic", freq = "all",
  date_range = "3y", from = NULL, to = Sys.Date(), print_step = 1L)
}
\description{
Returns
}
