% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd.R
\name{getmd}
\alias{getmd}
\title{get market data}
\usage{
getmd(symbol, frequency = "daily", from = "2010-01-01", to = Sys.time(),
  print_step = 1L, source = "yahoo", fillzero = FALSE)
}
\arguments{
\item{symbol}{symbols market data, which is available via getmd_symbol.}

\item{frequency}{default is daily. It supports daily, weekly and monthly for data from yahoo; daily for data from 163 and sina.}

\item{from}{the start date. Default is '2010-01-01'.}

\item{to}{the end date. Default is current system date.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}

\item{source}{data source. The available sources including 'yahoo', '163' and 'sina'.}

\item{fillzero}{logical. Defualt is FALSE If it is TRUE, the zeros in dataset will be filled with last non-zero values.}
}
\description{
\code{getmd} provides an interface to get historical data for different markets, such as forex, bond, money, stock and commodity, which are provided public sources.
}
\examples{
\dontrun{
# Example I
# get Chinese stock data from 163
dat = getmd(symbol=c('600000', '000001', '^000001', '^399001'), source="163")

dat2 = getmd(symbol=c('600000', '000001', '^000001', '^399001'), source="163", frequency="spot")

# get spot price of all A shares in sse and szse
dat3 = getmd(symbol='a', source="163", frequency="spot")

# get spot price of all index in sse and szse
dat4 = getmd(symbol='index', source="163", frequency="spot")


# Example II
# get Chinese commodity future price
dat_com1 = getmd(symbol='RB0', source="sina")

dat_com2 = getmd(symbol=c("AG0","AU0"), source="sina")


 # Example III
 # get data from yahoo
 dat_yahoo = getmd(symbol=c("^GSPC", "000001.SS", "EURUSD=X"))
 
 # for Chinese share
 dat_yahoo2 = getmd(c("000001", "^000001"))
}

}
