% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd_index_cons.R
\name{getmd_index_cons}
\alias{getmd_index_cons}
\title{get constituent of securities index}
\source{
\url{http://www.csindex.com.cn/zh-CN}
}
\usage{
getmd_index_cons(symbol)
}
\arguments{
\item{symbol}{the symbol of securities index. It supports Chinese securities index only at this moment.}
}
\description{
\code{getmd_index_cons} provides an interface to get the current constituent of securities index.
}
\examples{
\dontrun{
dt50 = getmd_index_cons("000016")

dt300 = getmd_index_cons("000300")

dt500 = getmd_index_cons("000905")
}

}
