% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmd.R
\name{getmd_symbol}
\alias{getmd_symbol}
\title{get symbols of market data}
\usage{
getmd_symbol(source = NULL, market = NULL)
}
\arguments{
\item{source}{data source. The available data sources including 'yahoo', '163' and 'sina'.}
}
\description{
\code{getmd_symbol} provides an interface to get symbols of different markets. yahoo (main currencies, commodities and world-indices), 163 (stock shares in sse and szse) and sina (commodity futures in dce, shfe, sge and zce).
}
\examples{
\dontrun{
syb_yahoo = getmd_symbol(source="yahoo")

syb_commodity_cn = getmd_symbol(source="sina")

syb_stock_cn = getmd_symbol(source="163")
}

}
