% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_0cate.R
\name{md_cate}
\alias{md_cate}
\title{query main market data by category}
\usage{
md_cate(cate = NULL, symbol = NULL, date_range = "3y", from = NULL,
  to = Sys.Date(), print_step = 1L, ...)
}
\arguments{
\item{cate}{the market category, forex, money, bond, index, commodity. Default is NULL.}

\item{symbol}{symbols of main market indicators.}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is '3y'.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{print_step}{a non-negative integer, which will print symbol name by each print_step iteration. Default is 1L.}

\item{...}{ignored parameters}
}
\description{
\code{md_cate} provides an interface to access main market data in five categories, including forex, money, bond, index, commodity.
}
\examples{
\donttest{
dat = md_cate()
}

}
