% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_future.R
\name{md_future_symbol}
\alias{md_future_symbol}
\title{symbol of future market data}
\usage{
md_future_symbol(...)
}
\arguments{
\item{...}{ignored parameters}
}
\description{
\code{md_future_symbol} returns all future symbols that provided by sina finance, see details on \url{http://vip.stock.finance.sina.com.cn/quotes_service/view/qihuohangqing.html} or \url{http://vip.stock.finance.sina.com.cn/mkt/#global_qh})
}
\examples{
\dontrun{
sybs = md_future_symbol()
}

}
