% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_stock_symbol.R
\name{md_stock_symbol}
\alias{md_stock_symbol}
\title{symbol components of exchange}
\usage{
md_stock_symbol(exchange = NULL)
}
\arguments{
\item{exchange}{the available stock exchanges are sse, szse, hkex, amex, nasdaq, nyse.}
}
\description{
\code{md_stock_symbol} returns all stock symbols by exchange
}
\examples{
\dontrun{
# get stock symbols in a stock exchange
## specify the exchanges
ex_syb1 = md_stock_symbol(exchange = c('sse', 'szse'))

## choose exchanges interactivly
ex_syb2 = md_stock_symbol()

}

}
