% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_addti.R
\name{ped_addti}
\alias{ped_addti}
\title{create technical indicators}
\usage{
ped_addti(dt, ti = list(sma = list(n = 20), sma = list(n = 50)), ...)
}
\arguments{
\item{dt}{time series datasets}

\item{ti}{list of technical indicators, overlay indicators include mm, sma, ema, smma, bb, sar, and oscillators indicators such as macd, roc, ppo, rsi, cci.
\enumerate{
\item overlay technical indicators
}

\itemize{
\item sma
\item ema
\item smma
\item bb
\item sar
}
\enumerate{
\item oscillator technical indicators
}

\itemize{
\item macd
\item ppo
\item rsi
\item cci
\item roc
}}

\item{...}{ignored}
}
\description{
\code{ped_addti} adds technical indicators to a dataset
}
\examples{
dt = md("^000001", source=163)

dt_ti = ped_addti(dt, ti=list(sma=list(n=20), sma=list(n=50)))

}
