% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_freqto.R
\name{ped_dailyto}
\alias{ped_dailyto}
\title{convert the frequency of daily data}
\usage{
ped_dailyto(dt, freq, print_step = 1L)
}
\arguments{
\item{dt}{time series datasets}

\item{freq}{the frequency input data will converted to It supports weekly, monthly, quarterly and yearly.}

\item{print_step}{A non-negative integer, which will print symbol name by each print_step iteration. Default is 1.}
}
\description{
convert the frequency of daily data
}
