% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_index.R
\name{ped_index}
\alias{ped_index}
\title{create an index}
\usage{
ped_index(dt, chain_index, weight, base_index = 1, base_date = "2010-01-01")
}
\arguments{
\item{dt}{input timeseries datasets}

\item{chain_index}{the name of chain index}

\item{weight}{the name of weight variable}

\item{base_index}{the base value of index, default is 1}

\item{base_date}{the date base of index, default is "2010-01-01"}
}
\description{
ped_index using the method of geometrically weighted averages to create an index based on multiple timeseries datasets.
}
\examples{
\dontrun{
banks_symbol = c("601988", "601288", "601398", "601939", "601328")

dat = getmd(banks_symbol, source="163", from="1900-01-01")

library(data.table)
dat = lapply(dat, function(x) x[, change_pct:=change_pct+1])

bankindex = ped_index(dat, "change_pct", "cap_total")
}

}
