% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_plot.R
\name{ped_plot}
\alias{ped_plot}
\title{create a chart for timeseries data}
\usage{
ped_plot(dt, y = "close", type = "line", title = NULL)
}
\arguments{
\item{dt}{a time series dataset}

\item{y}{the variable display on chart}

\item{type}{chart style, including line, candle, bar}

\item{title}{chart title}
}
\description{
`ped_plot` creates a charts for a time series dataset.
}
\examples{
\dontrun{
ssec = getmd("^000001", source="163")

p = ped_plot(ssec, title="SSEC")
print(p)
}

}
