% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_opr.R
\name{pq_opr}
\alias{pq_opr}
\title{dataframe operation}
\usage{
pq_opr(dt, opr, x = "close", rm_na = FALSE, ...)
}
\arguments{
\item{dt}{a list/dataframe of time series datasets.}

\item{opr}{operation string.}

\item{x}{the numeric column names, defaults to close.}

\item{rm_na}{weather to remove NA values when perform arithmetic.}

\item{...}{additional parameters.}
}
\description{
It performs arithmetic operation on numeric columns on multiple series.
}
\examples{
data("dt_banks")

dt1 = pq_opr(dt_banks, '601288.SS/601988.SS')
print(dt1)

dt2 = pq_opr(dt_banks, c('(601288.SS+601988.SS)/2', '(601288.SS*601988.SS)^0.5'))
print(dt2)

}
