% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_return.R
\name{pq_return}
\alias{pq_return}
\title{calculating returns by frequency}
\usage{
pq_return(dt, x, freq = "monthly", method = "arithmetic",
  rcol_name = "return", cols_keep = NULL, date_range = "max",
  from = NULL, to = Sys.Date(), print_step = 1L)
}
\arguments{
\item{dt}{a list/dataframe of daily series dataset}

\item{x}{the column name of adjusted asset price.}

\item{freq}{the frequency of returns. It supports c('all', 'daily', 'weekly', 'monthly', 'quarterly', 'yearly').}

\item{method}{the method to calculate returns.}

\item{rcol_name}{setting the name of return column, defaults to return.}

\item{cols_keep}{the columns keep in the return data. The columns of symbol, name and date will always kept if they are exist in the input data.}

\item{date_range}{date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.}

\item{from}{the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.}

\item{to}{the end date. Default is the current date.}

\item{print_step}{a non-negative integer. Print symbol name by each print_step iteration. Default is 1L.}
}
\description{
\code{pq_return} calculates returns for daily series based on specified column, frequency and method type.
}
\examples{
\donttest{
data(ssec)

# create a close_adj column
datadj = md_stock_adjust(ssec, adjust = FALSE)

# set freq
dts_returns1 = pq_return(datadj, x = 'close_adj', freq = 'all')

# set method
dts_returns2 = pq_return(datadj, x = 'close_adj', method = 'log')

# set cols_keep
dts_returns3 = pq_return(datadj, x = 'close_adj', cols_keep = 'cap_total')

}

}
